refactor: move `Remote...` and split `...Base`
